+++ src/lib.rs
@@ -21,6 +21,7 @@ pub mod digits;
 pub mod fs;
 #[cfg(feature = "image")]
 pub mod graphics;
+pub mod keymap;
 pub mod list;
 pub mod logging;
 pub mod mouse;